Fixed bugs related to turning off a channel.
When a channel was turned off, various problems arose.  Sometimes
the chart would stop running for all channels and could not be
restarted without program restart.  When the program was restarted
and scan started with the first channel turned off, an exception
was being thrown because necessary pointers were not being set for
masked or off channels.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>