Control board is now connected with by the same method as used by the UT boards: the host broadcasts a roll call on a UDP channel and listens for the boards to reply with their IP addresses.  Both Control boards and UT boards reply only to roll calls for their specific board type.
The UT boards used this method because they cannot read the chassis and
slot numbers until their FPGAs are programmed - thus they must make
contact with the host prior to being able to use those numbers to create
a known IP address.  After contact and FPGA load, the boards transmit their
chassis and slot numbers to the host so it can assign each board to the
appropriate channels.

The Control board ver 1.0 uses port D to access the chassis and slot
slot numbers.  On the Rabbit RCM4200, that port is setup as A/D inputs
by default - 8 tiny jumpers must be moved to use the port as digital I/O.
In addition, the Control board ver 1.0 itself must have 8 pullup resistors
added to those lines.  For now, systems only have one Control board so
it is adequate to wait for the answer from the first to respond - the
Control boards never do read their chassis and slot numbers.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>