Added code to fire Output 1 on the Control board when any signal crosses any threshold.  This was done for the Paragon Weldline #1 system.
The output is pulsed immediately after the incident.  Future systems will
need to be able to delay the marking as the marker is not usually located
right at the sensor.

Need to add code to assign thresholds to different outputs, disable in
Scan mode (maybe an option to turn on for Scan mode for use at Paragon
who only needs to run in Scan mode), and handle minor & major marking
for the minor and major thresholds.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>