Added code to allow for continuous marking mode where the marker is continuously fired while a threshold has been violated.
NOT TESTED - ABANDONDED - BETTER METHOD USED

Before this code was tested, it was decided it would be easier to fire
a marker pulse for each data point rather than turning the marker on and
then off.  This also reduces the chance that the marker will be left on
when the program is aborted abruptly.  The marker will effectively be
fired continuously for closely spaced violations.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>