Fixed bug where peaks were being missed while in SCAN mode if the speed was less than 10.  It had mistakenly been assumed that the peaks were being captured constantly, but it was actually only happening after the scan speed timer triggered.
Fixing this bug led to glitches in the plotting -- much was reorganized
regarding the plotting pointer and the data stuffing pointer.  Now, the
plotting pointer is always one behind the stuffing pointer so it doesn't
use data from a buffer slot still being overwritten with peak data.

There is now a bug -- when the screen is redrawn the traces shift 1 pixel
or so to the left.  Probably due to the fact that the buffer pointers now
start at the end of the buffer and the pixel pointers start at 0.

Signed-off-by: Mike <mike@folsomvillage.com>