Changed to load FPGA code in UT boards via UDP broadcasts instead of via TCP/IP.  This is much faster.
Need to add code to check if FPGA already loaded.  Currently, the TCP/IP
load is still in place, but it does nothing if it detects that the FPGA
was already loaded -- in this place by the preceding UDP code.  The UDP
code also needs to check for an already loaded FPGA and skip it if ALL
boards are already loaded.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>