Changed starting point of the data buffers to the beginning instead of the end.
Because at least one spot is skipped when the first data run is started,
the old code started at the end of the buffer so the first data would be
wrapped around and stored at the beginning.  This was overly complicated
to just save a single data spot.  The start of the piece is marked
properly regardless of where it starts in the buffer, so skipping one data
slot at the beginning is inconsequential.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>