Added a Demo Simulation control panel to control the simulated piece motion.  Added mechanism to transmit messages from control panel back to the simulator object to control the simulation.
NOTE: The buttons do not yet control anything.  The Forward button does
set a variable in the ControlSimulator class as a test of concept, but it
is not yet used.

In Hardware class, added precursor code to handle direction tracking of
the piece so that the traces can be reversed.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>