Belle Chasse file with trace per ducer, channels arranged as they will be used in production.
The ducer to leading edge of trace values in this file are those measured
by Mike Schoonover in IRNDT facility.  The values measured by the Tubo
personnel in Belle Chasse are probably more accurate, but each ducer was
measured from the ducer to the eye instead of to the leading edge of the
head.

The new values can be used to determine the distance between the ducers
and merged with this file to obtain better accuracy.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>