Fixed bug with DAC gains not being sent to the remotes by adding an override function calculateGateTimeLocation in DACGate which recalculates the gain each time the level is changed.
When the software gain is loaded from the cal file in Channel.java, it
is now passed on to the DACs by calling setSoftwareGain rather than
setting the softwareGain synced variable directly.

Changed getDataChanged method in SyncedVariable to getDataChangedFlag to
better reflect its purpose.

Changed return type from boolean to void for calculateGateTimeLocation
in BasicGate as the return value was no longer needed.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>