Fixed bug where copying DAC gates between channels caused them to be replicated -- all channels ended up pointing to the same DAC values.
The pointers to the synced variable objects were being copied instead of
the values in those objects being copied.

Also set such pointers to "final" so that they can never be changed
accidentally.

Changed the setValue method in all the Synced Variable classes so that
it only sets the value if the new value is different or if the update is
forced. This reduced a lot of duplicate code elsewhere.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>