Changed method of setting the rep rate registers in the remote fpga from direct write to calling a function in the Rabbit which then set the registers after checking the value for range.
Changed wait loop in fpga to include a sleep time instead of
relying on loop counter. The FPGA_LOAD_TIMEOUT was adjusted
to account for this to result in a two second timeout while
waiting on a packet request from the remote. This method is more
reliable across computing platforms as the timing will be
consistent.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>