Added code to set the Rabbit control byte. Created method prepareHPIARegistersForFPGAToDSPCommunications for clarity. Created variables and methods to track the number of DSP messages sent and the number of Ack packets received in return.
Removed bit definitions for DSP's flags1 as those functions have been
moved to a different variable in the DSP. Now the DSP does not alter any
bits in flags1 so that entire variable can be overwritten by the host
without conflict.

Added FLAW_WALL_MODE, WALL_MAP_MODE, RABBIT_FLAW_WALL_MODE,
RABBIT_WALL_MAP_MODE, BASIC_PEAK_COLLECTOR, WALL_MAPPER to select
different types of processing for the UTBoard object, the Rabbit, and
the DSP.

Renamed sendWallChannelFlag to setWallChannelFlag as the variable is not
actually sent to the DSPs -- it is only used in the host.

All methods which called sendBytes with command MESSAGE_DSP_CMD now call
sendMessageToDSP which counts the messages.

Added processAllAvailableDataPackets to clear out all data packets
in the buffer.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>