Changed hardwareDelay and softwareDelay to synchronized variables are they are changed by the user in the GUI thread and sent to the DSPs in a different thread. Added code to send the changes as necessary.
Changed calls to sendWallChannelFlag to setWallChannelFlag -- same method
but new name in UTBoard.

Added code to sendDataChangesToRemotes to check if DSPs sent back an Ack
packet for every message sent. NOTE: Sometimes still reports a mismatch
if the DSPs take a little time to respond. Increase wait time a bit?


Signed-off-by: Mike Schoonover <mike@folsomvillage.com>