Repaired what looks to be earlier mistakes regarding actions at the end of pipe (such as saving data) which seemed to work but were not proper.
Variable hdwVs.trackToEndOfPiece is now set false when the tracker
reaches 0 -- the prepareForNewPiece was being set true over and over.
Main was checking this flag and probably tried to save the segment over
and over.

A note added to check if the hokey prepareForNewPiece flag being set true
in collectEncoderDataInspectMode is working okay. Seems like it would be
finishing things off too early -- the ducers haven't reached the end of the
pipe when the photo eye clears. Even though they are lifted, they might
still need to track to the end of the pipe to make lengths kosher.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>