Added code to find boards configured to be map data sources and add them to WallMapDataSaver object.
Added temporary code to save data to text files -- needs to move to a
text format version of WallMapDataSaver.

Moved the configure method after init in Capulin1 as is done in other
classes.

Renamed getChassisSlotAddressFromRemote and
  getChassisSlotAddressOverrideFromFile in UTBoard to avoid confusion
with getters that return values from the UTBoard object itself.

UTBoards now load which onboard hardware channel provides map data
 -- currently only channel 0 is used but allowing this to be changed will
make it easier to use an existing configuration where wall channels are
not always channel 0.

UTBoards now load which map channel they are to provide data for.

Added method UTBoard::triggerMapAdvance to store a flag in the data buffer
when the map object is advanced so the data locations can be later
determined by scanning the data's control flags.

Removed UTBoard::saveDataBufferToTextFile -- temporarily moved to
WallMapDataSaverTuboBinary, later to be moved to a text format version
of that class lineage.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>