Program now records the start and stop inspection locations in each UT Board which is a map data source.
The code has been changed regarding map advancing -- it was advancing
map2D plotter object itself, based on its sensor offset in its UTBoard.
However, each map source sensor needed to be advanced, and often some
of those are not tied to a map. So now the sensor delay offsets drive
the advance trigger and the map follows along for whichever sensors it
is tied to.

Per the above paragraph, all code related to map delay offsets has now
been changed to use UTBoard.mapSensorDelay variables.

UTBoard.recordMapDataEnabled is not currently used -- in future will
control whether or not data from map packets is stored. I should not be
stored when the source sensors have not reached the test piece or after
they have left the end. It DOES need to be recorded before and after the
start/stop inspection events so their data/plotters will trace from the
start of the piece to the end.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>