Set up board slots 7 & 8 to be wall mapping data sources.
Set up board slot 7 to be the 2D wall map plotter data source. The plotter
only plots data from board 7. Data from board 8 is stored and used to
create the wall mapping output file along with slot 7 data.

Swapped channels 30 and 34 so that the Wall on channel 30 would be the
only channel on a board (slot 8) that is a wall mapping source. Currently
only one channel per board is allowed for wall mapping -- channel 29 Wall
was left on slot 7.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>