Added ability to drive an audible alarm via an ethernet relay module rather than by the Control board.
Refactored so "output pulse 1" and such are now "output audible
alarm" for clarity.

Need to add code to Control board to read which channel fires the
alarm from the config file as is done with the IO module class.

NOTE: Also made changes to Hardware.java, but there were other
changes related to the Wall Mapping so that file was not included
with this commit as it is more important to keep the Wall Mapping
changes in the same commit.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>