Fixed bug where causing a program restart due to changing jobs or loading presets was causing a freeze.
The new code added to access the IO modules installs a URL
factory handler. This can only be done once per JVM instance, so
the restart was crashing as it attempted to set it again.

Caught the Error (Error is thrown instead of an Exception) and
ignored the second attemp.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>