Fixed bug where UDP broadcasts were failing in systems with multiple network adapters, such as when an Internet connection is present along with the adapter used in the system.
The UDP connection uses whatever adapter the system chooses if not
explicitly bound to an adapter. This fix looks for the adapter
with the IP address 169.254.*.* which is that used by the system.

The 169.254.*.* address is the default assigned by Windows when no
DHCP server is found. In the future, if multiple adapters are to
be used with that default, further code will have to be added to
detect which is connected to the remotes.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>