Fixed two bugs which were preventing reverse and data/trace erasure.
Hardware.java

When the code at the end of moveEncoders was moved to moveTraces and
a call to moveMaps was added, an error was introduced by
"(pixelsMoved > 0)" -- the value is negative for reversing. This was
solved by removing the test altogther -- a few lines before the function
bails if pixelsMoved is zero, so reaching the bottom means that it is
non-zero and the moving functions need to be called.

PlotterData.java

If the trace was backed up more than one spot before the plotting thread
could react, the second erasure cleared all flags to zero for the first
erasure. As the extraction point would still on the first point, the
plotting thread would not see a data valid or a data erased flag and
would do nothing. Now, when erasing, any erase flag already in place is
left in place.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>