Fixed bug for XXtreme/IOS where min/max wall were not being printed properly on print-outs.
The first problem was that they had multiple charts named "Wall"
so the first chart so named, which was not actually a Wall chart,
was being used but failing since there was no "Min" or "Max"
trace.

The second problem was that the length of the tube was not
being calculated and stored when the Stop button was clicked.
For the IOS unit, the photo-eyes are not used so the Stop click
denoted the end of scan. Without a valid length, the program
could not scan the data for min/max wall values. This was fixed
mainly by adding code to Hardware.java and Capulin1.java.

The third problem was the math in ChartGroup.java used to
calculate the number of data points from the length.

The unit installation files were updated some time ago from the
unit's computer.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>