Removed old code and comments related to tracking to the end of the tube after the head clears the end of the tube.
Originally, the system expected the end-of-tube signal after the leading
vertical eye cleared the end, so a distance was set to continue tracking
until a specified distance after the eye cleared.

The final PLC software at Belle Chasse actually gave the end-of-tube
signal after the trailing vertical eye cleared so there was no need to
continue tracking after the eye cleared. This solved another problem --
the carriage could not continue turning the encoder all the way to home
end as the belt was lifted off the encoder wheel. It did turn until the
trailing vertical eye cleared, so using that eye as the end-of-tube trigger
worked.

A new class just for handling encoders and eyes needs to be created so
clean code can be written for each class specific to the encoder/eye
setup for each unit type.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>