The stopXloc stored in the map file is now set equal to the inspected distance -- from the first revolution at head down to the last at head up. This seems to work best with the Tubo Map viewer program.
The nXloc location value for each helix is now calculated using the
average helix. The average helix is now calculated from the actual
number of revolutions over the distance inspected. This is easier and
more accurate than using the position increment flags stored in the data
which are based on the encoder counts.

Corrected offsets between the ducers. Needs to be loaded from the config
file in the future.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>