Unit in Belle Chasse was not displaying minimum wall after inspection. The simulation on the laptop was displaying, but the search area was improper (between the end masks).
There was an error in calculating length as the new value stored is in
feet instead of inches.

Unknown when the change in length calculation was made.

Seems like the unit at Belle Chasse should have still worked but with the
limited search area. Might have been recording a value off the screen and
the square flag drawn off screen as well? It did show a value for the min.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>