When Stop button is clicked before an inspection is completed, the length of the run is set to whatever distance has been inspected -- ignoring the fact that the distance between the leading and trailing eyes should be subtracted. The trailing eye might not have been triggered yet, this length is simply an estimated value.
The value was being saved in inches where it is being saved as feet in
other parts of the code. Fixed this.

Removed comment about the value being used to determine if any inspection
data needs to be saved. That is actually done by Main.segmentStarted by
calling each trace to see if data has been added.

Signed-off-by: Mike Schoonover <mike@folsomvillage.com>