An earlier attempt to fix problem where signal spike was occuring at the end of DAC gates was a bad fix..this repairs that. (commit with tag BAD_FIX_SUPERCEDED_LATER)
It was thought that the spiking was caused by a round-off error gap
between the end of a DAC gate (start + width) and the start of the
following gate...so the earlier fix added one to the width. This actually
made the problem worse as it was really caused by OVERLAPPING of the gates.
This caused gain to be applied twice to the overlapped data.

Fixed her by not allowing the start of a gate overlap the previous gate...
the start of each gate is force to the start+width of the preceding gate.
This ony causes an adjustment by a few data points (15 nS).

Signed-off-by: Mike Schoonover <mike@skoonie.com>