Marker counts to marker now divided by 8 before sending to PLC...the PLC no longer divides by 8.
The PLC encoder counts are 1/8 of the computer's encoder counts as the
PLC is using the lower rollers with the encoders built into the motors.

The number of characters for the marker counts in the message to the PLC
is 5 digits. After changing the computer to the top rollers with the
Automation Direct encoders, the values became too large to fit into the
5 characters. Dividing by 8 in the computer solved the problem.

Signed-off-by: Mike Schoonover <mike@skoonie.com>